requirements.txt

crewai==0.13.0
litellm==1.10.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
anthropic==0.5.0
google-cloud-aiplatform==1.36.4